chore: restructure methods to various files to improve readability
